cron

+++ src/models/Pagos.js
@@ -17,9 +17,10 @@ const Pagos = sequelize.define("pagos", {
     allowNull: false,
   },
   valorDepositado: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.DECIMAL(10, 2),
     allowNull: false,
   },
+
   entidad: {
     type: DataTypes.STRING,
     allowNull: true,
